+++ CHANGELOG.md
@@ -113,3 +113,4 @@
 - Fix #109: Update error message in `DivTest` for `dataAttributes()` method assertion (@terabytesoftw)
 - Fix #110: Remove suffix `Element` in classes names for consistency across the codebase (@terabytesoftw)
 - Fix #111: Update documentation to specify valid tags for `Block` and `Inline` enum class references (@terabytesoftw)
+- Fix #112: Simplify attributes in `BaseInline` class and improve `Span` element with specific traits and update related tests (@terabytesoftw)